// datanode-ui/src/libs/hooks/useSystemSettings/useSystemSettings.enums.ts
/*
 *
 * Copyright 2023 Odysseus Data Services, Inc.
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
 */

export enum UseSystemSettings {
  GET_LIST = "GET_LIST",
  GET_LIST_DONE = "GET_LIST_DONE",
  GET_LIST_FAILED = "GET_LIST_FAILED",
  UPDATED_VALUE = "UPDATED_VALUE",
  UPDATED_VALUE_DONE = "UPDATED_VALUE_DONE",
  UPDATED_VALUE_FAILED = "UPDATED_VALUE_FAILED"
}